Fix float precision
